requirements.txt
requests==2.31.0
python-dotenv>=0.19.0
aiohttp==3.9.1
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
pyyaml==6.0.1
tqdm==4.66.1
tenacity==8.2.3
watchdog==3.0.0 
# Required dependencies
rich>=10.0.0
click>=8.0.0
rich>=13.9.4    
setuptools>=75.6.0

# Optional dependencies (for enhanced code quality checks)
mypy>=1.0.0  # optional: for type checking
pylint>=2.17.0  # optional: for code quality analysis
pyyaml>=6.0
click>=8.0.0
astor>=0.8.1
tenacity>=8.0.0

# Testing dependencies
pytest-cov>=3.0.0